Made servo and motor and button classes. Lots of testing.

+++ src/org/usfirst/frc/team2635/robot/SingleMotorTest.java
@@ -12,7 +12,7 @@ public class SingleMotorTest implements IInput<Double>
 	@Override
 	public void setInput(Double input)
 	{
-		motor.set(input);
+		motor.set((double)input);
 	}
 
 }
